spit off OTCrypto symmetric key functions into interface opentxs::CryptoSymmetric
